Check NEXT_PUBLIC_VERCEL_ENV for platform metadata on Vercel config
We noticed a bug that all client-side logs from our Vercel deployments
are being sent with the metadata field vercel.env set to "production".
This line looks to be the cause -- the Vercel config is only checking
for the VERCEL_ENV environment variable, which isn't available in client
components - so it falls back to the NODE_ENV which is "production" for both Vercel
preview and production deployments. The corresponding client variable is NEXT_PUBLIC_VERCEL_ENV,
so this PR changes the metadata to check for either
NEXT_PUBLIC_VERCEL_ENV or VERCEL_ENV before falling back to NODE_ENV.